# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: سند – منصة الخدمات والمهنيين في اليمن
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /conversations/{id}/messages:
    get:
      operationId: getMessages
      tags: [messages]
      summary: Get messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  # ─── NOTIFICATIONS ───────────────────────────────────────────────────────────
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [name, phone, password, role]
      properties:
        name:
          type: string
        phone:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [client, provider]
        categoryId:
          type: ["integer", "null"]
        city:
          type: ["string", "null"]
        district:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        yearsExperience:
          type: ["integer", "null"]
        hourlyRate:
          type: ["number", "null"]
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
    LoginInput:
      type: object
      required: [phone, password]
      properties:
        phone:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, name, phone, role, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [client, provider, admin]
        status:
          type: string
          enum: [active, pending, banned]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string

    # ── CATEGORY ────────────────────────────────────────────────────────────────
...
    Message:
      type: object
      required: [id, conversationId, senderId, content, createdAt]
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        senderId:
          type: integer
        senderName:
          type: string
        content:
          type: string
        imageUrl:
          type: ["string", "null"]
        isRead:
          type: boolean
        createdAt:
          type: string
    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
        imageUrl:
          type: ["string", "null"]

    # ── NOTIFICATION ─────────────────────────────────────────────────────────────
...