# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * سند – منصة الخدمات والمهنيين في اليمن
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';
import type { UserStatus } from './userStatus';

export interface User {
  id: number;
  name: string;
  phone: string;
  role: UserRole;
  status: UserStatus;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
}